# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for User Service",
        "title": "API Documentation",
        "contact": {},
        "version": "0.2"
    },
    "paths": {
...
        "/numerical-method/interpolation/linear-newton": {
            "post": {
                "description": "Create the linear newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linear Newton"
                ],
                "summary": "Create Linear Newton",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqLinearNewton"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.LinearNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/linear-newton/{id}": {
            "get": {
                "description": "Get the linear newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Linear Newton"
                ],
                "summary": "Get Linear Newton",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LinearNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/polynomial-newton": {
            "post": {
                "description": "Create the polynomial newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polynomial Newton"
                ],
                "summary": "Create Polynomial Newton",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqPolynomialNewton"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PolynomialNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/polynomial-newton/{id}": {
            "get": {
                "description": "Get the polynomial newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Polynomial Newton"
                ],
                "summary": "Get Polynomial Newton",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PolynomialNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-lagrange": {
            "post": {
                "description": "Create the quadratic lagrange data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Lagrange"
                ],
                "summary": "Create Quadratic Lagrange",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqQuadraticLagrange"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticLagrange"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-lagrange/{id}": {
            "get": {
                "description": "Get the quadratic lagrange data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Lagrange"
                ],
                "summary": "Get Quadratic Lagrange",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticLagrange"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-newton": {
            "post": {
                "description": "Create the quadratic newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Newton"
                ],
                "summary": "Create Quadratic Newton",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqQuadraticNewton"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-newton/{id}": {
            "get": {
                "description": "Get the quadratic newton data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Newton"
                ],
                "summary": "Get Quadratic Newton",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticNewton"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-spline": {
            "post": {
                "description": "Create the quadratic spline data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Spline"
                ],
                "summary": "Create Quadratic Spline",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqQuadraticSpline"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticSpline"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/interpolation/quadratic-spline/{id}": {
            "get": {
                "description": "Get the quadratic spline data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quadratic Spline"
                ],
                "summary": "Get Quadratic Spline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QuadraticSpline"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LinearNewton": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "models.PolynomialNewton": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "models.QuadraticLagrange": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
        "models.QuadraticNewton": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
        "models.QuadraticSpline": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validations.ReqLinearNewton": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "validations.ReqPolynomialNewton": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
...
        "validations.ReqQuadraticLagrange": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
        "validations.ReqQuadraticNewton": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "number"
                }
            }
        },
        "validations.ReqQuadraticSpline": {
            "type": "object",
            "properties": {
                "point": {
                    "type": "string"
                },
                "points": {
                    "type": "string"
                },
                "xvalue": {
                    "type": "string"
                }
            }
        },
...
    }
}